Update ic-vetkd-utils to the latest version
