+++ motoko/encrypted-notes-dapp-vetkd/vetkd_user_lib/README.md
@@ -3,7 +3,7 @@ The two files in this directory were generated by running
 ```
 wasm-pack build --target web --release
 ```
-in the package root of a clone of the git repository containing the vetKD user library (https://github.com/dfinity-lab/vetkd/vetkd_user_lib). Note that `wasm-pack` outputs the
+in the package root of a clone of the git repository containing the vetKD user library (https://github.com/dfinity/ic/tree/master/packages/ic-vetkd-utils). Note that `wasm-pack` outputs the
 files into a `pkg/` folder.
 
 See also https://github.com/rollup/plugins/tree/master/packages/wasm#using-with-wasm-bindgen-and-wasm-pack.
